[tvOS] added SpeakerDetailViewController
